Utility: introduce allocateAligned().
I chose allocateAligned() over alignedAllocation() because it's two
letters shorter and a verb, and alignedAlloc[ate]() feels like broken
English to me, don't like that either.

Also it's a new dedicated header because of the Array include and
because this header may get fancy new things like magic ring buffers in
the future.